% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletForest-package.R
\name{DirichletRF-package}
\alias{DirichletRF-package}
\title{DirichletRF: Dirichlet Random Forest for Compositional Data}
\description{
This package implements Dirichlet Random Forest for modeling and predicting
compositional data using maximum likelihood estimation or method of moments.
}
