% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_forest.R
\name{DirichletRF}
\alias{DirichletRF}
\title{Build a Dirichlet Random Forest for Compositional Responses}
\usage{
DirichletRF(
  X,
  Y,
  num.trees = 100,
  max.depth = 10,
  min.node.size = 5,
  mtry = -1,
  seed = 123,
  est.method = "mom",
  num.cores = -1
)
}
\arguments{
\item{X}{A numeric (n x p) matrix of covariates. Note that the current
version only allows numeric covariates. Users may use one-hot encoding
to possibly include categorical covariates.}

\item{Y}{A numeric (n x k) matrix of compositional responses. Each row
should sum to 1. That is, data should already be normalised if needed.}

\item{num.trees}{Number of trees grown in the forest. Default is 100.}

\item{max.depth}{Maximum depth of trees. Default is 10.}

\item{min.node.size}{Minimum size of observations in each tree leaf.
Default is 5. Note that nodes with sizes smaller than
\code{min.node.size} can occur.}

\item{mtry}{Number of covariates randomly selected as candidates at each
split. Default is \code{sqrt(p)}, indicated by \code{-1}.}

\item{seed}{The seed of the C++ random number generator.}

\item{est.method}{Parameter estimation method for the Dirichlet
distribution when splitting is done. Users may either use maximum
likelihood (\code{"mle"}) or method of moments (\code{"mom"}).
Default is \code{"mom"}.}

\item{num.cores}{Number of OpenMP threads used for parallel tree building.
The default is \code{-1} which uses all the cores on the system minus 1.
Users may also specify \code{1} which means that the forest will be
built sequentially.}
}
\value{
A list of the forest which contains the following:
\describe{
\item{\code{type}}{Parallelisation type used: \code{"openmp"} or
\code{"sequential"}.}
\item{\code{num.cores}}{Number of cores used.}
\item{\code{num.trees}}{Total number of trees in the forest.}
\item{\code{Y_train}}{Training data used (no OOB).}
\item{\code{fitted}}{A list of fitted values:
\describe{
\item{\code{alpha_hat}}{Estimated Dirichlet alpha parameters
(n x k matrix).}
\item{\code{mean_based}}{Mean-based fitted values (n x k matrix).}
\item{\code{param_based}}{Parameter-based fitted values obtained
by normalising \code{alpha_hat} (n x k matrix).}
}
}
\item{\code{residuals}}{A list of residuals (Y - fitted):
\describe{
\item{\code{mean_based}}{Residuals from mean-based predictions.}
\item{\code{param_based}}{Residuals from parameter-based
predictions.}
}
}
}
}
\description{
Build a Dirichlet random forest for compositional responses. In
compositional data analysis (CoDA), parts reside in the simplex, and this
random forest ensures model output abide by CoDA principles. The
implementation uses OpenMP for parallel tree building. Note that this implementation
does not support out-of-bag (OOB) error estimation.
}
\details{
The forest provides two types of fitted values: mean-based predictions
(derived from sample means at each leaf) and parameter-based predictions
(derived from normalised Dirichlet alpha parameters).
}
\examples{
# Small toy example (auto-tested)
set.seed(42)
n <- 50; p <- 2
X <- matrix(rnorm(n * p), n, p)
G <- matrix(rgamma(n * 3, shape = rep(c(2, 3, 4), each = n)), n, 3)
Y <- G / rowSums(G)
forest <- DirichletRF(X, Y, num.trees = 5, num.cores = 1)
print(forest)
Xtest <- matrix(rnorm(5 * p), 5, p)
pred  <- predict(forest, Xtest)

\donttest{
# Larger example
n <- 500; p <- 4
X <- matrix(rnorm(n * p), n, p)
alpha <- c(2, 3, 4)
G <- matrix(rgamma(n * length(alpha), shape = rep(alpha, each = n)),
            n, length(alpha))
Y <- G / rowSums(G)

forest1 <- DirichletRF(X, Y, num.trees = 100, num.cores = 1)
forest2 <- DirichletRF(X, Y, num.trees = 100)

alpha_hat   <- forest1$fitted$alpha_hat
mean_fit    <- forest1$fitted$mean_based
param_fit   <- forest1$fitted$param_based
resid_mean  <- forest1$residuals$mean_based
resid_param <- forest1$residuals$param_based

Xtest <- matrix(rnorm(10 * p), 10, p)
pred  <- predict(forest1, Xtest)
param_pred <- pred$alpha_predictions / rowSums(pred$alpha_predictions)
}

}
\references{
Masoumifard, K., van der Westhuizen, S., & Gardner-Lubbe, S. (In press).
Dirichlet-random forest for predicting compositional data. In A. Bekker,
J. Ferreira, & P. Nagar (Eds.), \emph{Environmental Statistics: Innovative
Methods and Applications}. CRC Press.
}
\seealso{
\code{\link{predict.dirichlet_forest}} for making predictions on new data.
}
