% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_forest.R
\name{predict.dirichlet_forest}
\alias{predict.dirichlet_forest}
\title{Predict with a Dirichlet Forest}
\usage{
\method{predict}{dirichlet_forest}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{dirichlet_forest} object.}

\item{newdata}{A numeric matrix of new covariates (n_new x p).}

\item{...}{Currently unused.}
}
\value{
A list with the following elements:
\describe{
\item{\code{alpha_predictions}}{Estimated Dirichlet alpha parameters
for each new observation (n_new x k matrix).}
\item{\code{mean_predictions}}{Mean-based compositional predictions
(n_new x k matrix).}
}
}
\description{
Makes predictions using a fitted \code{dirichlet_forest} object returned
by \code{\link{DirichletRF}}.
}
\examples{
# Small toy example (auto-tested)
set.seed(42)
n <- 50; p <- 2
X <- matrix(rnorm(n * p), n, p)
G <- matrix(rgamma(n * 3, shape = rep(c(2, 3, 4), each = n)), n, 3)
Y <- G / rowSums(G)
forest  <- DirichletRF(X, Y, num.trees = 5, num.cores = 1)
Xtest   <- matrix(rnorm(5 * p), 5, p)
pred    <- predict(forest, Xtest)
pred$mean_predictions

\donttest{
n <- 500; p <- 4
X <- matrix(rnorm(n * p), n, p)
alpha <- c(2, 3, 4)
G <- matrix(rgamma(n * length(alpha), shape = rep(alpha, each = n)),
            n, length(alpha))
Y <- G / rowSums(G)
forest <- DirichletRF(X, Y, num.trees = 50, num.cores = 1)
Xtest  <- matrix(rnorm(10 * p), 10, p)
pred   <- predict(forest, Xtest)
param_pred  <- pred$alpha_predictions / rowSums(pred$alpha_predictions)
single_pred <- predict(forest, Xtest[1, , drop = FALSE])
}
}
